requirements.txt
scikit-learn
joblib
mlflow==2.7.1
numpy
azure-ai-ml # For MLClient
azure-identity # For DefaultAzureCredential
azureml-mlflow # For MLflow integration with Azure ML
